fix: fix temp user guard
