+++ src/app/(user)/layout.tsx
@@ -4,19 +4,16 @@ import type { Metadata } from "next";
 import { Header } from "@/components/common/Header";
 import { Footer } from "@/components/common/Footer";
 import classes from "./layout.module.css";
-import { TempUserGuard } from "@/components/common/Auth/AuthModule";
 export const metadata: Metadata = {
   title: "S-TOP 기술교류회",
   description: "S-TOP 기술교류회",
 };
 
-export default async function UserLayout({
+export default function UserLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  await TempUserGuard();
-
   return (
     <>
       <Header />
